Development

+++ src/main/java/com/collabera/interview/enroll/service/EnrollServiceImpl.java
@@ -37,25 +37,49 @@ public class EnrollServiceImpl implements EnrollService{
 
     @Override
     public void deleteEnroll(String id) {
-        this.enrollRepository.deleteById(id);
+        if (enrollRepository.existsById(id)) {
+            this.enrollRepository.deleteById(id);
+        }else {
+            throw new EnrollNotFoundException(format("Enrollee with id %s not found ", id));
+        }
     }
 
     @Override
-    public void saveDependents(String enrollId, List<Dependent> dependentList) {
-        Enroll enroll = enrollRepository.findById(enrollId).get();
-        enroll.getDependants().addAll(dependentList);
+    public Enroll saveDependents(String enrollId, List<Dependent> dependentList) {
+        Enroll enroll = enrollRepository.findById(enrollId).orElseThrow(()->new EnrollNotFoundException(format("Enrollee with id %s not found", enrollId)));
+        enroll.getDependents().addAll(dependentList);
+        return enroll;
     }
 
     @Override
     public void deleteDependentByEnrollId(String dependentId, String enrollId) {
-        Enroll enroll = enrollRepository.findById(enrollId).get();
-        enroll.getDependants().remove(enroll);
+        Enroll enroll = enrollRepository.findById(enrollId).orElseThrow(()->new EnrollNotFoundException(format("Enrollee with id %s not found", enrollId)));
+       List<Dependent>dependentList= enroll.getDependents();
+       if (!dependentList.isEmpty()){
+           dependentList.remove(dependentList.stream()
+                   .filter(dependent -> dependent.getId().equals(dependentId)).findFirst().orElseThrow(()->new EnrollNotFoundException(format("dependent with id %s not found", dependentId))));
+           enroll.setDependents(dependentList);
+           enrollRepository.save(enroll);
+       }
+
     }
 
     @Override
-    public void modifyDependent(Dependent dependent, String enrollId) {
-        Enroll enroll = enrollRepository.findById(enrollId).get();
-         enroll.getDependants().add(dependent);
+    public Enroll modifyDependent(Dependent dependent, String enrollId) {
+        Enroll enroll = enrollRepository.findById(enrollId).orElseThrow(()->new EnrollNotFoundException(format("Enrollee with id %s not found", enrollId)));
+        Dependent dependent1= enroll.getDependents().stream()
+                .filter(dependent2 -> dependent2.getId().equals(dependent.getId()))
+                .findFirst()
+                .orElseThrow(()->new EnrollNotFoundException(format("dependent  with id %s not found", dependent.getId())));
+        int dependentIndex= enroll.getDependents().indexOf(dependent1);
+        dependent1.setDateOfBirth(dependent.getDateOfBirth());
+        dependent1.setName(dependent.getName());
+   List<Dependent>dependentList=    enroll.getDependents();
+   dependentList.set(dependentIndex, dependent1);
+   enroll.setDependents(dependentList);
+        enrollRepository.save(enroll);
+        return enroll;
+
     }
